bug fix: marshalling a bitblock into a specialized posit<32,2> used incorrect cast

+++ posit/specialized/posit_32_2.hpp
@@ -144,7 +144,7 @@ namespace sw {
 		explicit operator unsigned int() const { return to_int(); }
 
 		posit& set(sw::unum::bitblock<NBITS_IS_32>& raw) {
-			_bits = uint8_t(raw.to_ulong());
+			_bits = uint32_t(raw.to_ulong());
 			return *this;
 		}
 		posit& set_raw_bits(uint64_t value) {
